==> code/2_download_singularity_CI_data.ipynb <==
# import project modules
import download_data
# import packages
import pandas as pd
from os import path

# full BA list
ba_list = ['EIA.CISO', 'EIA.ISNE', 'EIA.BPAT','EIA.MISO', 'EIA.NYIS', 'EIA.PJM', 'EIA.SWPP',
'AEC', 'AECI', 'AVA', 'AZPS', 'BANC', 'CHPD', 'CPLE', 'CPLW', 'DOPD', 'DUK', 'EPE', 'ERCO', 'FMPP', 
'FPC', 'FPL', 'GCPD', 'GVL', 'HST', 'IID', 'IPCO', 'JEA', 'LDWP', 'LGEE', 'NEVP', 'NSB', 
'NWMT', 'PACE', 'PACW', 'PGE', 'PNM', 'PSCO', 'PSEI', 'SC', 'SCEG', 'SCL', 'SEC', 
'SOCO', 'SPA', 'SRP', 'TAL', 'TEC', 'TEPC', 'TIDC', 'TPWR', 'TVA', 'WACM', 'WALC', 'WAUW']

# we want 2019 data
start_date = '2019-01-01'
end_date = '2019-12-31'

# set overwrite = True if you want to overwrite already downloaded files
overwrite = True

# Enter your Singularity API password
api_credentials = ''



for ba in ba_list:

    if path.exists(f'../data/processed/singularity_efs/{ba}.csv'):
        if overwrite == False:
            print(f'{ba} already downloaded.')
        elif overwrite == True:
            # download the data
            df = download_data.download_singularity_data(
                api_key=api_credentials, ba=ba, start_date=start_date, end_date=end_date)

            # save the data
            df.to_csv(f'../data/processed/singularity_efs/{ba}.csv')
    else:
        # download the data
        df = download_data.download_singularity_data(
            api_key=api_credentials, ba=ba, start_date=start_date, end_date=end_date)

        # save the data
        df.to_csv(f'../data/processed/singularity_efs/{ba}.csv')
